Refactor: move build_error_dict into the Analyser call, since it will be backend specific
