+++ django_migration_linter/sql_analyser/utils.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import re
 from copy import deepcopy
 
 
@@ -34,22 +33,3 @@ def update_migration_tests(base_tests, specific_tests):
         for key in override_test.keys():
             migration_test_dict[key] = override_test[key]
     return base_tests
-
-
-def build_error_dict(migration_test, sql_statement):
-    table_search = (
-        re.search("TABLE `([^`]*)`", sql_statement, re.IGNORECASE)
-        if isinstance(sql_statement, str)
-        else None
-    )
-    col_search = (
-        re.search("COLUMN `([^`]*)`", sql_statement, re.IGNORECASE)
-        if isinstance(sql_statement, str)
-        else None
-    )
-    return {
-        "err_msg": migration_test["err_msg"],
-        "code": migration_test["code"],
-        "table": table_search.group(1) if table_search else None,
-        "column": col_search.group(1) if col_search else None,
-    }
